+++ src/test/java/com/quasiris/qsf/pipeline/filter/elastic/Elastic2SearchResultMappingTransformerTest.java
@@ -2,6 +2,7 @@ package com.quasiris.qsf.pipeline.filter.elastic;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.quasiris.qsf.pipeline.filter.elastic.bean.ElasticResult;
+import com.quasiris.qsf.response.Document;
 import com.quasiris.qsf.response.SearchResult;
 import org.junit.jupiter.api.Test;
 
@@ -27,6 +28,20 @@ class Elastic2SearchResultMappingTransformerTest {
         assertEquals("Features", firstSnippet.get("title"));
         assertEquals(0.0, firstSnippet.get("_score"));
         assertEquals(0, firstSnippet.get("_offset"));
+
+    }
+    @Test
+    void transformInnerhitsCollapse() throws Exception {
+        Elastic2SearchResultMappingTransformer transformer = new Elastic2SearchResultMappingTransformer();
+        ElasticResult elasticResult = readElasticResultFromFile("innerhits-collapse.json");
+        SearchResult searchResult = transformer.transform(elasticResult);
+
+        List<Document> productTypeGrouped = searchResult.getDocuments().get(0).getChildDocument("product_type_grouped");
+        assertEquals(5, productTypeGrouped.size());
+        assertEquals("iphone Handy Hülle blau title-id-doc-02", productTypeGrouped.get(1).getFieldValue("title"));
+
+
+        System.out.println(searchResult);
 
     }